chore: move traits into axes
